# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  controllers.RejectRequest:
    properties:
      catatan:
        type: string
    type: object
...
  entities.User:
    properties:
      email:
        type: string
      foto:
        type: string
      id:
        type: integer
      nama:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
  description: Dokumentasi API backend Sistem Informasi Karang Taruna
  title: API Sistem Informasi Karang Taruna
  version: "1.0"
paths:
...
  /api/kegiatan/{id}/foto:
    post:
      consumes:
      - multipart/form-data
      description: Upload foto kegiatan setelah kegiatan disetujui
      parameters:
      - description: ID Kegiatan
        in: path
        name: id
        required: true
        type: integer
      - description: File foto kegiatan
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Upload foto kegiatan
      tags:
      - Ketua Divisi Kegiatan
...
  /api/ketua/kegiatan/{id}/approve:
    patch:
      description: Menyetujui pengajuan kegiatan
      parameters:
      - description: ID Kegiatan
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Approve kegiatan
      tags:
      - Ketua Umum Kegiatan
  /api/ketua/kegiatan/{id}/reject:
    patch:
      consumes:
      - application/json
      description: Menolak pengajuan kegiatan disertai catatan
      parameters:
      - description: ID Kegiatan
        in: path
        name: id
        required: true
        type: integer
      - description: Catatan penolakan
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RejectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Reject kegiatan
      tags:
      - Ketua Umum Kegiatan
...
  /kegiatan/{id}/foto:
    get:
      description: Mengambil daftar foto kegiatan berdasarkan ID kegiatan
      parameters:
      - description: ID Kegiatan
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get foto kegiatan
      tags:
      - Public Kegiatan
  /login:
    post:
      consumes:
      - application/json
      description: Login untuk ketua divisi atau ketua umum
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      description: Mendaftarkan user dengan role ketua_divisi
      parameters:
      - description: Data register
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register ketua divisi
      tags:
      - Auth
  /register-ketua:
    post:
      consumes:
      - application/json
      description: Mendaftarkan user dengan role ketua_umum
      parameters:
      - description: Data register ketua umum
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register ketua umum
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"